Pause and resume running collectors around fork
