+++ lib/vernier/fork.rb
@@ -4,10 +4,15 @@ module Vernier
   if ::Process.respond_to?(:_fork)
     module ForkHooks
       def _fork
+        running_collectors = ObjectSpace.each_object(Vernier::Collector).select(&:running?)
+        running_collectors.each(&:pause)
         pid = super
-        if pid == 0 # We're in the child
-          ObjectSpace.each_object(Vernier::Collector) do |collector|
-             Vernier.cancel_profile
+        if pid == 0
+          # We're in the child
+        else
+          # We're in the parent
+          running_collectors.each do |collector|
+            collector.resume
           end
         end
         pid
